# backend/docs/openapi.yaml
components:
...
  securitySchemes:
    apiKeyAuth:
      description: API key for device authentication. Provide the API key as a Bearer
        token.
      in: header
      name: Authorization
      type: apiKey
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  contact:
    name: MOTO Support
  description: API for the MOTO school management system
  title: MOTO API
  version: 1.0.0
openapi: 3.0.3
paths:
...
  /api/groups/{id}/substitutions:
    get:
      description: Generated from routes
      parameters:
      - description: id parameter
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: GET /api/groups/{id}/substitutions
      tags:
      - Groups
...
  /api/staff/{id}/substitutions:
    get:
      description: Generated from routes
      parameters:
      - description: id parameter
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: GET /api/staff/{id}/substitutions
      tags:
      - Staff
...
  /api/substitutions:
    get:
      description: Generated from routes
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: GET /api/substitutions
      tags:
      - Substitutions
    post:
      description: Generated from routes
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: POST /api/substitutions
      tags:
      - Substitutions
  /api/substitutions/{id}:
    delete:
      description: Generated from routes
      parameters:
      - description: id parameter
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: DELETE /api/substitutions/{id}
      tags:
      - Substitutions
    get:
      description: Generated from routes
      parameters:
      - description: id parameter
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: GET /api/substitutions/{id}
      tags:
      - Substitutions
    put:
      description: Generated from routes
      parameters:
      - description: id parameter
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: PUT /api/substitutions/{id}
      tags:
      - Substitutions
  /api/substitutions/active:
    get:
      description: Generated from routes
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: GET /api/substitutions/active
      tags:
      - Substitutions
...
servers:
- description: API Base URL
  url: /api